requirements.txt
Brotli==1.1.0
cachetools==4.2.4
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.4
dataclasses==0.6
google-api-core==1.34.0
google-api-python-client==2.6.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.1
google-drive==0.5.0
google-oauth==1.0.1
googleapis-common-protos==1.60.0
httplib2==0.22.0
idna==3.4
inflate64==0.3.1
multivolumefile==0.2.3
oauthlib==3.2.2
pip-autoremove==0.10.0
protobuf==3.20.3
psutil==5.9.5
py7zr==0.20.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybcj==1.0.1
pycparser==2.21
pycryptodomex==3.19.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyppmd==1.0.0
python-dotenv==1.0.0
pyzstd==0.15.9
rarfile==4.1
redis==5.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
spotipy==2.23.0
texttable==1.6.7
uritemplate==3.0.1
urllib3==2.0.5
